pyproject.toml
[project]
name = "event-driven-microservices"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.245.0",
    "constructs>=10.6.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
